pyproject.toml
[project]
name = "magnet-pow-miner"
version = "1.0.0"
requires-python = ">=3.8"
dependencies = [
    "python-dotenv>=1.0.0",
    "web3==6.15.1",
    "eth-account==0.11.1",
    "eth-typing==4.2.0",
    "pycryptodomex==3.22.0",
    "colorlog==6.9.0",
    "eth-hash==0.7.1",
    "keyboard==0.13.5",
    "numpy==2.2.5"
    
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"